requirements.txt
absl-py==2.0.0
aenum==3.1.15
affine==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
alabaster==0.7.13
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.3
bokeh==3.3.2
Brlapi==0.8.3
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.1
camelot-py==0.11.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.7
constantly==15.1.0
contourpy==1.1.1
croniter==1.4.1
cryptography==42.0.5
cssselect==1.2.0
cupshelpers==1.0
cycler==0.10.0
dask==2023.12.0
dataclasses-json==0.6.3
dateutils==0.6.12
dbus-python==1.2.18
decorator==5.1.1
deepdiff==6.6.0
defer==1.0.6
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.7
distributed==2023.12.0
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.4
django-allauth==0.61.1
django-email-verification==0.3.3
django-phonenumber-field==7.3.0
django-use-email-as-username==1.4.0
djangorestframework==3.15.1
docutils==0.20.1
duplicity==0.8.21
earthengine-api==0.1.384
Elixir==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.103.2
fasteners==0.14.1
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.1
Flask-Login==0.6.3
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.2
gast==0.5.4
gevent==24.2.1
geventhttpclient==2.2.1
ghostscript==0.7
google-api-core==2.15.0
google-api-python-client==2.112.0
google-auth==2.23.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.59.0
gyp==0.1
h11==0.14.0
h5py==3.10.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.6
imageio==2.33.1
imagesize==1.4.1
importlib-metadata==7.0.0
imutils==0.5.4
incremental==22.10.0
iniconfig==2.0.0
inquirer==3.1.3
ipython==8.16.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
JPype1==1.4.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.14.0
keyring==23.5.0
kiwisolver==1.4.5
labelImg==1.8.6
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lightgbm==4.2.0
lightning==2.0.9.post0
lightning-api-access==0.0.5
lightning-cloud==0.5.39
lightning-utilities==0.9.0
locket==1.0.0
lockfile==0.12.2
locust==2.27.0
louis==3.20.0
lxml==4.9.3
lz4==4.3.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
mysqlclient==2.1.1
netifaces==0.11.0
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.46
opencv-python==4.8.1.78
opencv-python-headless==4.5.5.64
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.1.1
paramiko==2.9.3
parsel==1.8.1
parso==0.8.3
partd==1.4.1
pbr==6.0.0
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pluggy==1.5.0
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.1.1
pydantic_core==2.4.0
PyDispatcher==2.0.7
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
pymongo==3.13.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pypdf==3.17.1
PyPDF2==3.0.1
pypdfium2==4.24.0
pypi==2.1
pyproject-toml==0.0.10
pyRFC3339==1.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.2.0
pytest-django==4.8.0
pytest-flask==1.3.0
pytest-mock==3.14.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-editor==1.0.4
python-multipart==0.0.6
python3-openid==3.2.0
pytorch-lightning==2.0.9.post0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==26.0.3
queuelib==1.6.2
rasterio==1.3.9
readchar==4.0.5
referencing==0.34.0
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.6.0
rpds-py==0.18.0
rsa==4.9
s2cloudless==1.7.2
scikit-learn==1.3.2
scipy==1.11.3
Scrapy==2.11.0
scrapy-djangoitem==1.1.1
scrapy-pipeline-mongodb==0.0.7
scrapy-user-agents==0.1.1
seaborn==0.13.0
SecretStorage==3.3.1
selenium==4.15.2
sentinelhub==3.10.1
service-identity==23.1.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
starsessions==1.3.0
supervision==0.15.0
sympy==1.12
systemd-python==234
tabula-py==2.9.0
tabulate==0.9.0
tblib==3.0.0
Tempita==0.5.2
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
tldextract==5.0.0
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
toolz==0.12.0
torch==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
tornado==6.4
tqdm==4.66.1
traitlets==5.11.2
trio==0.23.1
trio-websocket==0.11.1
triton==2.1.0
Twisted==22.10.0
txmongo==23.0.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
ua-parser==0.18.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ultralytics==8.0.196
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.1
usb-creator==0.3.7
user-agents==2.2.0
utm==0.7.0
uvicorn==0.23.2
validators==0.27.0
vboxapi==1.0
w3lib==2.1.2
wadllib==1.3.6
wcwidth==0.2.8
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.0
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==0.7.2
xyzservices==2023.10.1
yarl==1.9.2
zict==3.0.0
zipp==1.0.0
zope.event==5.0
zope.interface==6.1
